Key by :name and :term when saving to the DB
Currently, IDs are only generated for members with a member page.
This means that members without a page are not assigned IDs.

If we key by :id when saving, we would need to generate an ID for
members without a page. As member IDs are currently generated from
the member page URL, we would need to find another way of
generating IDs for members without pages. But, if a member without
a page were to eventually get a page, they would end up with a
new ID generated from their page's URL and thus appear to be a new
member.

Without keying by ID there should be no need to generate an ID for
every member. I believe the import process will assign an ID to
anyone without an ID and would use existing IDs where members
have them.