Ensure correct URL is captured for each member row
A row can contain none, one, or two anchor tags. When one or more links
are present, the URL of the member's individual page is always the one
containing text that matches their sort name. If no anchors contain
the matching text, then the member does not have an individual page.